fix: 리스폰스 수정, s3에 기존에 있는지 확인, content-type 수정

+++ core/src/main/java/com/project/configs/S3Manager.java
@@ -6,18 +6,23 @@ import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
+import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
+import com.amazonaws.util.IOUtils;
 import com.project.utils.DateUtils;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.annotation.PostConstruct;
+import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.util.Date;
 import java.util.UUID;
 
 @Component
+@RequiredArgsConstructor
 public class S3Manager {
     private AmazonS3 s3Client;
 
@@ -27,7 +32,7 @@ public class S3Manager {
     @Value("${cloud.aws.credentials.secretKey}")
     private String secretKey;
 
-    @Value("${cloud.aws.s3.bucket}")
+    @Value("${cloud.aws.bucket}")
     private String bucketName;
 
     @Value("${cloud.aws.region.static}")
@@ -45,14 +50,20 @@ public class S3Manager {
     public String uploadFile(MultipartFile file, String userName, String exerciseName) throws IOException {
         String createdAt = DateUtils.parseDateToString(new Date());
         String fileName = String.format("%s_%s_%s_%s", userName, exerciseName, createdAt, UUID.randomUUID().toString());
+        byte[] bytes = IOUtils.toByteArray(file.getInputStream());
+        ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream(bytes);
 
+        ObjectMetadata metadata = new ObjectMetadata();
+        metadata.setContentLength(bytes.length);
+        metadata.setContentType("video/mp4");
 
-        s3Client.putObject(new PutObjectRequest(bucketName, fileName, file.getInputStream(), null)
+        s3Client.putObject(new PutObjectRequest(bucketName, fileName, byteArrayInputStream, metadata)
                 .withCannedAcl(CannedAccessControlList.PublicRead));
-        return fileName;
+        return s3Client.getUrl(bucketName, fileName).toString();
     }
 
     public void deleteFile(String url) {
-        s3Client.deleteObject(bucketName, url);
+        if (s3Client.doesObjectExist(bucketName, url))
+            s3Client.deleteObject(bucketName, url);
     }
 }
